Changed back to ints, fixed ASTDef and environments

+++ BASE-0/ASTNode.java
@@ -1,6 +1,6 @@
 public interface ASTNode {
 
-    double eval(Environment e);
+    int eval(Environment e);
 	
     void compile(CodeBlock c, Environment e);
 }
